✨ 사진작가 리스트 url parms 추가
