+++ src/components/search/FilteringBox.js
@@ -1,9 +1,12 @@
 import styled from "styled-components";
+import { useNavigate, useLocation } from "react-router-dom";
 import { React, useState, useEffect } from "react";
 import CustomCalendar from "./CustomCalendar";
 import { regions, orders, category } from "./FilteringList.js";
 
 const FilteringBox = ({ onSearch, setIsFilteringOpen }) => {
+  const navigate = useNavigate();
+
   const [selectedRegion, setSelectedRegion] = useState("서울");
   const [selectedSubRegion, setSelectedSubRegion] = useState(null);
   const [selectedSection, setSelectedSection] = useState(null);
@@ -27,7 +30,28 @@ const FilteringBox = ({ onSearch, setIsFilteringOpen }) => {
   };
 
   const handleSearchBtnClick = (isFilteringOpen) => {
-    onSearch(selectedSubRegion, selectedSection, selectedDate, selectedOrder);
+    let endpoint = "/photographers";
+    const queryParams = [];
+    if (selectedSubRegion) {
+      queryParams.push(`areaId=${selectedSubRegion}`);
+    }
+    if (selectedSection) {
+      queryParams.push(`special=${selectedSection}`);
+    }
+    if (selectedDate) {
+      queryParams.push(`ableDate=${selectedDate}`);
+    }
+    if (selectedOrder && selectedOrder.length > 0) {
+      queryParams.push(`sort=${selectedOrder}`);
+    }
+
+    if (queryParams.length > 0) {
+      endpoint += "?" + queryParams.join("&");
+    }
+    console.log(endpoint);
+    navigate(endpoint);
+
+    // onSearch(selectedSubRegion, selectedSection, selectedDate, selectedOrder);
     setIsFilteringOpen(!isFilteringOpen);
   };
